Fixes
Fix clang warnings.
Fix usage calculation bug.

+++ src/analyzer.h
@@ -8,7 +8,7 @@ typedef struct analyzer_params_t {
   queue_t *out_queue;
   queue_t *watchdog_queue;
   queue_t *logger_queue;
-  int nprocs;
+  size_t nprocs;
 } analyzer_params_t;
 
 void *analyzer_thread(void *analyzer_params);
